some improvements
