+++ src/test/resources/application-connectedtest.properties
@@ -1,3 +1,5 @@
+datarepo.testWithDatabase=true
+
 # ct for connected test
 ct.ingestBucket=jade-testdata
 ct.nonDefaultRegionIngestBucket=jade-testdata-useastregion
