+++ src/test/resources/application-integrationtest.properties
@@ -1,4 +1,4 @@
-tdr.testNoDatabase=true
+datarepo.testWithDatabase=false
 
 # it for integration test
 it.jadeApiUrl=https://jade.datarepo-integration.broadinstitute.org
